📚 Adding documentation about configuring rodeo logger
Yes, we could have a DerivativeRodeo initializer...but we leverage the
Rodeo by way of IiifPrint.  So this makes sense as to the place to
configure these things.

In addition, I figured I'd share one of the things I did to help in the
debugging of the DerivativeRodeo integration.  Which has me thinking
that perhaps we should look at doing this with the IiifPrint gem.  After
all, debugging that is also challenging.

Related to:

- https://github.com/scientist-softserv/derivative_rodeo/issues/56