🚧 Revert using DerivativeRodeo
This commit reverts using the Derivative Rodeo for PDF splitting and
derivative generation.  It is also breadcrumbs for how to restore those
functions.  We revert from the Derivative Rodeo to the already
established IIIF Print pluggable derivatives derived from the Newspaper
works gem.

The reason to revert is that this branch includes several changes that
went into local testing of the DerivativeRodeo; and I want to capture
those wins and merge in an already long-running branch, to reduce the
chance of further branch drift.

For reference, local testing of the DerivativeRodeo has worked both with
and without having SpaceStone data for both PDF splitting and generating
derivatives (e.g.  thumbnails, word coordinates, alto files, and plain
text).  However, I had only done localized testing and I believe more
testing is warranted; namely how does the full text search work.

To consider is how we will:

- Test on staging with the Rodeo but not have it in play for Production

But that is an exercise for the person undoing this commit :)

Related to:

- https://github.com/scientist-softserv/derivative_rodeo/issues/56
- https://github.com/scientist-softserv/adventist-dl/issues/500