🧹 Conditional enabling of S3 for pre-processing
The default assumption of IIIF Print was S3 for pre-processing.  That
does not hold well for local development.  So I'm adding a configuration
to say "Use S3 if we have credentials" Otherwise don't.  This will
probably move to IIIF Print; but right now we have a slow build on IIIF
Print.

Related to:

- https://github.com/scientist-softserv/adventist-dl/pull/671

Co-authored-by: Shana Moore <shana@scientist.com>