🎁 Add feature flipper for PDF.js viewer
This commit will add a feature flipper to switch between using PDF.js or
Universal Viewer for displaying PDFs.  At the moment we have no
mechanism to turn off IIIF Print PDF processing, however, that was not
indicated as a requirement at this time so that is not included until we
have a need for it.