♻️ Re-arrange method structure in hopes of not swallowing exception
We are observing that this job is raising an exception, which is then
swallowed (and "pooped out" to the mailboxer notifications for the
submitting user; which is not helpful when tracking down the
particulars).

Confounding part is that the reported errors, don't seem to be affecting
the attachment of files.  In other words, we get a notification of
failure but it worked (the majority of the time).

Co-authored-by: LaRita Robinson <larita@scientist.com>