🐛 Check for pdf extension
This commit will add a check for the file set's extension (label)
instead of solely relying on the mime type. This will allow us to know
if it's a PDF (as best as we can anyway) even if CharacterizeJob did not
run.

Also, downgraded the version of simple_form because it was causing
crashes in development.