🎁 Filter certain file sets from manifests
This commit will remove thumbnail file sets and non-image file sets at a
much higher level than before.  Previously, we were filtering out the
thumbnails at the manifest builder level which is right before the
manifest gets generated.  This caused issues with serverless IIIF
resources when the file sets (such as thumbnails or text files) were not
properly ingested.  This shouldn't break the manifest generation since
these kinds of file sets are not able to be displayed in the Universal
Viewer.  Filtering these out at a higher level means it doesn't even
make it to the serverless IIIF portion of the code in IIIF Print.

Ref:
  - https://github.com/scientist-softserv/adventist-dl/issues/679