🎁 Only run thumbnail derivatives on thumbnails
We were seeing that our thumbnail files were getting word coordinates
derivatives, which is a waste of processing time.  The jobs would
eventually fail but ideally we don't want them to run at all for
thumbnails.

This commit will upgrade to the latest version of IIIF Print where a new
configurable class attribute has been added to enable a more granular
control over which derivatives are run on what files.

Also in this commit, we were seeing a nilClass error for a logger in
development, so that has also been fixed.

Ref:
  - https://github.com/scientist-softserv/adventist-dl/issues/676
  - https://github.com/scientist-softserv/iiif_print/pull/306