🐛 Fix for UV when characterize job fails
The problem here is that we are trying to render download options for
non-image files in the UV.  The original logic rejects anything that is
an image so PDFs, TXT, etc can be downloaded through the UV.  However,
if the characterize job fails then we have a `nil` for the mime type
which is why the error is thrown.  This commit will rework the logic to
account for that.

Ref:
  - https://github.com/scientist-softserv/adventist-dl/issues/679