add auxiliary worker to run different queues
Add a second worker that runs all GoodJob queues. The "base" worker will
run all queues except the :auxiliary queue.

This, combined with the high priority assigned to all jobs in the
:auxiliary queue, effectively builds a fast lane; all jobs run in the
:auxiliary queue will be executed before all other jobs by the auxiliary
worker.

The fact that the base worker does not run any jobs in the :auxiliary
queue means we can have this fast lane without putting every other job
on hold for, potentially, long periods of time.

P.S. The reason we use `before_enqueue` in ApplicationJob is because
GoodJob does not support setting a default priority per queue.